LayoutTests/http/tests/media/media-source/webm/video-media-source-add-and-remove-ids-expected.txt

Tests webkitSourceAddId() & webkitSourceRemoveId() methods


running testAddIdFailureCases
Got expected exception Error: INVALID_STATE_ERR: DOM Exception 11
EVENT(loadstart)
EVENT(webkitsourceopen)
Test empty ID case
Got expected exception Error: INVALID_ACCESS_ERR: DOM Exception 15
Test adding the same ID again.
Got expected exception Error: INVALID_STATE_ERR: DOM Exception 11
Test empty type.
Got expected exception Error: INVALID_ACCESS_ERR: DOM Exception 15
Test an unsupported type.
Got expected exception Error: NOT_SUPPORTED_ERR: DOM Exception 9
Test reaching sourceID limit.
Got expected exception Error: INVALID_STATE_ERR: DOM Exception 11

running testRemoveIdFailureCases
EVENT(loadstart)
EVENT(webkitsourceopen)
Test empty ID case
Got expected exception Error: INVALID_ACCESS_ERR: DOM Exception 15
Test removing an ID that was never added.
Got expected exception Error: SYNTAX_ERR: DOM Exception 12
Got expected exception Error: SYNTAX_ERR: DOM Exception 12
Test adding the same ID again.
Test that an ID can be removed while in the ended state.

running testAppendFailureCases
EVENT(loadstart)
EVENT(webkitsourceopen)
Test append with empty ID.
Got expected exception Error: INVALID_ACCESS_ERR: DOM Exception 15
Test append with an invalid ID
Got expected exception Error: SYNTAX_ERR: DOM Exception 12
Test append with a null buffer.
Got expected exception Error: INVALID_ACCESS_ERR: DOM Exception 15
Test a successful append.
Got expected exception Error: SYNTAX_ERR: DOM Exception 12
END OF TEST
